auto_email/requirements.txt
python-dotenv>=1.0.0
msal>=1.28.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
